pyproject.toml
[tool.poetry]
name = "codyssey-team-a"
version = "0.1.0"
description = "Backend API using FastAPI, PostgreSQL, and Redis"
authors = ["haram Jeong <haram001021@gmail.com>"]
readme = "README.md"
package-mode = false

packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"

fastapi = "^0.118.0"
uvicorn = {extras = ["standard"], version = "^0.37.0"}

python-dotenv = "^1.1.1"
pydantic-settings = "^2.11.0"

sqlalchemy = {extras = ["asyncio"], version = "^2.0.43"}

asyncpg = "^0.29.0"
greenlet = "^3.0.3"

geoalchemy2 = "^0.15.2"
alembic = "^1.16.5"

redis = "^6.4.0"
shapely = "^2.0.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"